changed layout

+++ server.js
@@ -17,7 +17,7 @@ const app = express();
 
 app.use( express.static( 'public'))
 
-app.use( cors() );
+// app.use( cors() );
 
 app.use(express.json())
 
@@ -100,8 +100,8 @@ app.delete("/remove-image/:image", async (req, resp)=>{
 app.delete("/remove-files", async (req, resp)=>{
 
 
-    let imageDir = `./public/portfolio/images`;
-    let mediaDir = `./public/portfolio/media`; // without backslash
+    let imageDir = path.join(__dirname, `/public/portfolio/images`);
+    let mediaDir = path.join(__dirname, `/public/portfolio/media`); // without backslash
 
     const imageFiles = await fs.readdirSync(imageDir);
     const mediaFiles = await fs.readdirSync(mediaDir);
@@ -109,7 +109,7 @@ app.delete("/remove-files", async (req, resp)=>{
     // remove files from image folder
     for( const file of imageFiles ){
 
-        if( file !== "default.jpeg" ){
+        if( file && file !== "default.jpeg" ){
 
             fs.unlinkSync(`./public/portfolio/images/${file}`);
 
@@ -118,7 +118,19 @@ app.delete("/remove-files", async (req, resp)=>{
     }
 
     // remove all files from /media
-    await fsExtra.emptyDirSync(mediaDir);
+    for( const file of mediaFiles ){
+
+        if( file ){
+
+            fs.unlinkSync(`./public/portfolio/media/${file}`);
+
+        }
+
+    }
+
+
+    // remove all files from /media
+    // await fsExtra.emptyDirSync(mediaDir);
 
 
     resp.status(200).send("files removed");
